refactor: :recycle: NAV_ROUTES -> useNavRoutes
