+++ app/src/components/elements/NavBar/NavMenu.tsx
@@ -1,9 +1,11 @@
 import { Spacer, VStack } from '@/components/common';
-import { NAV_ROUTES } from '@/routes/NAV_ROUTES';
 import { NavItem } from './NavItem';
+import { useNavRoutes } from './hooks';
 
 // TODO: SidebarMenu와 TabBar의 로직 유사하므로 통합 필요
 export const NavMenu = () => {
+  const { NAV_ROUTES } = useNavRoutes();
+
   return (
     <VStack as="ul" w="100%" h="100%" spacing="2rem">
       <NavItem route={NAV_ROUTES.HOME} />
