+++ app/src/components/elements/NavBar/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useNavRoutes';
