--- app/src/components/elements/TabBar/index.ts
@@ -1,1 +1,0 @@
-export * from './TabBar';
