+++ app/src/routes/ROUTES.ts
@@ -1,6 +1,6 @@
 export const ROUTES = {
   ROOT: '/',
-  PROFILE: '/profile/:username',
+  PROFILE: '/profile/:userId',
   PROJECT: '/project/:projectName',
   LEADERBOARD: '/leaderboard',
   EVALLOG: '/evallog',
